requirements.txt
dj-database-url>=0.5.0
Django>=3.0.3
gunicorn>=20.0.4
pytz>=2019.3
sqlparse>=0.3.0
psycopg2-binary>=2.8.4
whitenoise>=5.0.1

djangorestframework~=3.11.0
